+++ Changes.md
@@ -12,6 +12,7 @@ Change History
 * Added configuration options for the default count values for "/w pre" and "/w post".
 * Added paging of "/w file list" output.
 * Added paging of "/w ore" output.
+* Added a configuration option for the Watson command prefix.
 
 1.4.7 (2013-01-23)
 ------------------